+++ AdventOfCode/2024/Day01/Day01.cs
@@ -4,7 +4,7 @@ public class Day01 : Day
 {
     public override bool PrintTime { get => false; set { } }
     
-    public override int CalculatePart01(string kindOfInput, string pathPrefix)
+    public override long CalculatePart01(string kindOfInput, string pathPrefix)
     {
         StreamReader streamReader = new StreamReader(pathPrefix + "\\" + kindOfInput + ".txt");
         string? currentLine = streamReader.ReadLine(); 
@@ -32,7 +32,7 @@ public class Day01 : Day
         return answerValue;
     }
 
-    public override int CalculatePart02(string kindOfInput, string pathPrefix)
+    public override long CalculatePart02(string kindOfInput, string pathPrefix)
     {
         StreamReader streamReader = new StreamReader(pathPrefix + "\\" + kindOfInput + ".txt");
         string? currentLine = streamReader.ReadLine();
